+++ lib/virtualCrypto_web/views/oauth2/authorize_view.ex
@@ -1,0 +1,5 @@
+defmodule VirtualCryptoWeb.OAuth2.AuthorizeView do
+  use VirtualCryptoWeb, :view
+  # authorize.html.eex
+  import Phoenix.Controller, only: [get_csrf_token: 0]
+end
